+++ src/rule.js
@@ -10,19 +10,23 @@ define(function(require, exports, module) {
         initialize: function(name, operator) {
             this.name = name;
 
-                if (operator instanceof RegExp) {
-                    this.operator =  function(opts, commit) {
-                        var result = operator.test($(opts.element).val());
+            if (operator instanceof RegExp) {
+                this.operator =  function(opts, commit) {
+                    var result = operator.test($(opts.element).val());
+                    commit(result ? null : opts.rule, _getMsg(opts, result));
+                };
+            } else if (typeof operator == 'function') {
+                this.operator = function(opts, commit) {
+                    var result = operator(opts, function(result, msg) {
                         commit(result ? null : opts.rule, _getMsg(opts, result));
-                    };
-                } else if (typeof operator == 'function') {
-                    this.operator = function(opts, commit) {
-                        var result = operator(opts, commit);
-                        if (result !== undefined)
-                            commit(result ? null : opts.rule, _getMsg(opts, result));
-                    };
-                } else
-                    throw new Error('The second argument must be a regexp or a function.');
+                    });
+                    if (result !== undefined) {
+                        commit(result ? null : opts.rule, _getMsg(opts, result));
+                    }
+                };
+            } else {
+                throw new Error('The second argument must be a regexp or a function.');
+            }
         },
 
         and: function(name, options) {
@@ -45,7 +49,7 @@ define(function(require, exports, module) {
                         target.operator(opts, commit);
                     }
                 });
-            }
+            };
 
             return new Rule(null, operator);
         },
@@ -70,7 +74,7 @@ define(function(require, exports, module) {
                         commit(null, _getMsg(opts, true));
                     }
                 });
-            }
+            };
 
             return new Rule(null, operator);
         },