+++ src/validator.js
@@ -48,7 +48,9 @@ define(function(require, exports, module) {
         addItem: function(cfg) {
             Validator.superclass.addItem.apply(this, [].slice.call(arguments));
             var item = this.query(cfg.element);
-            this._saveExplainMessage(item);
+            if (item) {
+              this._saveExplainMessage(item);
+            }
             return this;
         },