+++ HISTORY.md
@@ -6,7 +6,7 @@
 
 *   **重要**：itemValidated 事件原来接受的参数是 `(element, err)` 现在改变为 `(err, results, element)`
 *   addRule、addItem 和 setMessage 支持一次操作多个。且都返回 this 对象，支持链式操作。
-*   支持初始化时传入非 form 外层容器作为 element。
+*   支持初始化时传入**非** form 外层容器作为 element。
 *   destroy 的时候增加恢复 novalidate 属性值。
 
 ## 0.8.8